blueprints/industry/healthcare-triage/backend/requirements.txt
flask
flask-cors
google-generativeai
chromadb
sentence-transformers
cryptography
python-dotenv
requests
numpy
pandas
scikit-learn
nltk